CDL-1688: re-factored functional tests for error messages i18n
