+++ functionaltests/src/test/resources/specs/response/I18nPayoneResponse.html
@@ -1,0 +1,146 @@
+<html xmlns:c="http://www.concordion.org/2007/concordion">
+<head>
+    <title>I18nPayoneResponseFixture</title>
+    <meta http-equiv="Content-Type" content="text/html; charset=UTF-8"/>
+</head>
+<body>
+<h1>Payone responses localization (internationalization) test</h1>
+<p>
+    Test that Payone service returns error messages in user language, not only English or other default.
+</p>
+
+<div class="example">
+    <p>
+    <ul>
+        <li>payment.paymentMethodInfo.method: <span c:set="#creditCardpaymentMethod">CREDIT_CARD</span></li>
+        <li>payment.transactions[0].type: <span c:set="#transactionType">AUTHORIZATION</span></li>
+    </ul>
+    Given the following payments for <strong>Credit Card</strong>:
+    </p>
+    <table c:execute="#paymentId = createCardPayment(#paymentName, #creditCardpaymentMethod, #transactionType, #centAmount, #currencyCode, #languageCode)">
+        <tr>
+            <th c:set="#paymentName">Payment Name</th>
+            <th c:set="#centAmount">Cent Amount</th>
+            <th c:set="#currencyCode">Currency Code</th>
+            <th c:set="#languageCode">Language Code</th>
+            <th c:echo="#paymentId">Payment ID (for information only)</th>
+        </tr>
+        <tr>
+            <td>Payment 1</td>
+            <td>0</td>
+            <td>USD</td>
+            <td>de</td>
+            <td></td>
+        </tr>
+        <tr>
+            <td>Payment 2</td>
+            <td>0</td>
+            <td>EUR</td>
+            <td>en</td>
+            <td></td>
+        </tr>
+        <tr>
+            <td>Payment 3</td>
+            <td>0</td>
+            <td>USD</td>
+            <td>nl</td>
+            <td></td>
+        </tr>
+        <tr>
+            <td>Payment 4</td>
+            <td>0</td>
+            <td>EUR</td>
+            <td>es</td>
+            <td></td>
+        </tr>
+        <tr>
+            <td>Payment 5</td>
+            <td>0</td>
+            <td>EUR</td>
+            <td></td><!-- With empty language we expect to have error message in English -->
+            <td></td>
+        </tr>
+        <tr>
+            <td>Payment 6</td>
+            <td>0</td>
+            <td>EUR</td>
+            <td>zz</td><!-- With wrong language we expect to have error message in English -->
+            <td></td>
+        </tr>
+    </table>
+
+    <hr/>
+
+    <p><h3>Testing response error messages localization</h3></p>
+
+    <p>
+        Read following values from <strong>ERROR</strong> JSON responses:
+    <ul>
+        <li><span c:set="#status">status</span></li>
+        <li><span c:set="#errorcode">errorcode</span></li>
+        <li><span c:set="#errormessage">errormessage</span></li>
+        <li><span c:set="#customermessage">customermessage</span></li>
+    </ul>
+    </p>
+    <table c:execute="#response = handleErrorJsonResponse(#paymentName)">
+        <tr>
+            <th c:set="#paymentName">Payment Name</th>
+            <th c:assertEquals="#response.jsonNode.get(#status).asText()">Status</th>
+            <th c:assertEquals="#response.jsonNode.get(#errorcode).asText()">Error Code</th>
+            <th c:assertEquals="#response.jsonNode.get(#errormessage).asText()">Error Message</th>
+            <th c:assertEquals="#response.jsonNode.get(#customermessage).asText()">Customer Error (localised)</th>
+            <th c:echo="#response.languageCode">Language</th>
+        </tr>
+        <tr>
+            <td>Payment 1</td>
+            <td>ERROR</td>
+            <td>1016</td>
+            <td>Parameter {amount} faulty or missing</td>
+            <td>Bei der Bearbeitung dieser Transaktion ist ein Fehler aufgetreten (Falsche Parameter).</td>
+            <td></td>
+        </tr>
+        <tr>
+            <td>Payment 2</td>
+            <td>ERROR</td>
+            <td>1016</td>
+            <td>Parameter {amount} faulty or missing</td>
+            <td>An error occured while processing this transaction (wrong parameters).</td>
+            <td></td>
+        </tr>
+        <tr>
+            <td>Payment 3</td>
+            <td>ERROR</td>
+            <td>1016</td>
+            <td>Parameter {amount} faulty or missing</td>
+            <td>Er is bij de bewerking van deze transactie een fout opgetreden (verkeerde parameters).</td>
+            <td></td>
+        </tr>
+        <tr>
+            <td>Payment 4</td>
+            <td>ERROR</td>
+            <td>1016</td>
+            <td>Parameter {amount} faulty or missing</td>
+            <td>Surgió un error durante el procesamiento de esta transacción. (Parámetros erróneos)</td>
+            <td></td>
+        </tr>
+        <tr>
+            <td>Payment 5</td>
+            <td>ERROR</td>
+            <td>1016</td>
+            <td>Parameter {amount} faulty or missing</td>
+            <td>An error occured while processing this transaction (wrong parameters).</td>
+            <td></td>
+        </tr>
+        <tr>
+            <td>Payment 6</td>
+            <td>ERROR</td>
+            <td>1341</td>
+            <td>Parameter {language} faulty or missing</td>
+            <td>Invalid language. Please verify your data.</td>
+            <td></td>
+        </tr>
+    </table>
+</div>
+
+</body>
+</html>
